# server/lumincoin_backend.openapi.yaml
openapi: 3.0.3
info:
  title: Backend for JS project
  version: 1.0.0
  contact: {}
servers:
  - url: http://localhost
  - url: localhost
paths:
  /api/refresh:
    post:
      tags:
        - Auth
      summary: Refresh
      description: Refresh
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  example: '{{refreshToken}}'
            examples:
              Refresh:
                value:
                  refreshToken: '{{refreshToken}}'
      responses:
        '200':
          description: ''
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
...